+++ test/mocks/obsidian.ts
@@ -217,6 +217,12 @@ export class PluginSettingTab {
     }
 }
 
+export class FileSystemAdapter {}
+
+export function normalizePath(path: string): string {
+    return path.replace(/\\/g, '/').replace(/\/{2,}/g, '/').replace(/^\/|\/$/g, '');
+}
+
 export class Setting {
     constructor(_containerEl: HTMLElement) {}